feat: make server side pages with client side components
