+++ src/app/analytics/page.module.css
@@ -1,18 +1,5 @@
-.chartDiv {
+.main {
     width: 100%;
-    display: flex;
     justify-content: center;
-}
-
-.loadingContainer {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100vw;
-    height: 100vh;
     display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: rgba(255, 255, 255, 0.8);
-    z-index: 1000;
 }
