+++ src/app/new-report/page.module.css
@@ -57,16 +57,4 @@
     background-color: var(--purple-hover);
   }
   
-  
-.loadingContainer {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: rgba(255, 255, 255, 0.8);
-  z-index: 1000;
-}
+
