+++ src/app/new/page.module.css
@@ -27,15 +27,3 @@
     height: 80%;
 }
 
-.loadingContainer {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100vw;
-    height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: rgba(255, 255, 255, 0.8);
-    z-index: 1000;
-}
