+++ src/components/PatientDetails/PatientDetails.module.css
@@ -1,0 +1,44 @@
+.container{
+    height: 95%;display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+}
+
+.chartDiv {
+    display: flex;
+    justify-content: center;
+    margin: 0 auto;
+    width: 80%;
+    cursor: pointer;
+
+}
+
+.infosDiv {
+    background-color: var(--super-light-purple);
+    border-radius: var(--border-radius);
+    width: 100%;
+    margin: 0 auto;
+    padding: 1rem 2rem;
+    border: 1px solid var(--gray);
+}
+
+.patientName {
+    overflow: hidden;
+    white-space: nowrap;
+    text-overflow: ellipsis;
+    max-width: 90%;
+}
+
+
+@media only screen and (max-width: 768px) {
+    .chartDiv {
+        height: 100%;
+        width: 95%;
+    }
+
+    .infosDiv {
+        width: 95%;
+    }
+
+
+}
