+++ src/styles/colors.css
@@ -1,11 +1,11 @@
-:root{
-    --white: #FAFAFA;
-    --super-light-gray: #efefef00;
-    --light-gray: #E2E1E1;
-    --gray: #CECBCB;
-    --purple: #3E0649;
-    --purple-hover: #3e0649;
-    --teal: #388B8B;
-    --box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.2);
-    --border-radius: 4rem;
-  } 
+:root {
+  --white: #FAFAFA;
+  --super-light-gray: #efefef00;
+  --super-light-purple: #cfc2eb33;
+  --gray: #CECBCB;
+  --purple: #3E0649;
+  --purple-hover: #3e0649;
+  --teal: #388B8B;
+  --box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.2);
+  --border-radius: 4rem;
+}
